Transposed unfollow file from JS to TS
